Adds readme, fixes test runs & first movements
First of all, the required README.md is added. It contains a quick
overview of what we try to achieve, what we use and how to use it.

Next, a small missing piece in our build script gets fixed. Tests can
run via Gradle now.

Last but definitely not least, a rover can be initialised and moved
within a given frame. Well, as long as the direction is EAST that is.
More is coming